+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Smarter Subcommanders doing nothing for viewers when Queller was the allied AI
 - Guardians failing AI mods of the "load" type
 - Prevent creation of invalid duplicate unit maps
+- Cluster enemy behaviour would change to reflect tech it didn't have if the player had subcommanders
 
 ## v6.2.1 - 2026-07-17